+++ asdf/extension/_legacy.py
@@ -1,5 +1,4 @@
 import abc
-from functools import lru_cache
 
 from .. import resolver, types
 from ..type_index import AsdfTypeIndex
@@ -150,39 +149,6 @@ class AsdfExtensionList:
         return self._validators
 
 
-def get_cached_asdf_extension_list(extensions):
-    """
-    Get a previously created AsdfExtensionList for the specified
-    extensions, or create and cache one if necessary.  Building
-    the type index is expensive, so it helps performance to reuse
-    the index when possible.
-
-    Parameters
-    ----------
-    extensions : list of asdf.extension.AsdfExtension
-
-    Returns
-    -------
-    asdf.extension.AsdfExtensionList
-    """
-    from ._extension import ExtensionProxy
-
-    # The tuple makes the extensions hashable so that we
-    # can pass them to the lru_cache method.  The ExtensionProxy
-    # overrides __hash__ to return the hashed object id of the wrapped
-    # extension, so this will method will only return the same
-    # AsdfExtensionList if the list contains identical extension
-    # instances in identical order.
-    extensions = tuple(ExtensionProxy.maybe_wrap(e) for e in extensions)
-
-    return _get_cached_asdf_extension_list(extensions)
-
-
-@lru_cache()
-def _get_cached_asdf_extension_list(extensions):
-    return AsdfExtensionList(extensions)
-
-
 # A kludge in asdf.util.get_class_name allows this class to retain
 # its original name, despite being moved from extension.py to
 # this file.
@@ -215,7 +181,9 @@ class _DefaultExtensions:
 
     @property
     def extension_list(self):
-        return get_cached_asdf_extension_list(self.extensions)
+        from ..config import get_config
+
+        return get_config().extension_list
 
     @property
     def package_metadata(self):